Update: S3 에서 이미지 삭제시 null 처리 추가

+++ src/main/java/com/chone/server/domains/s3/service/S3Service.java
@@ -46,6 +46,10 @@ public class S3Service {
 
   public void removeFile(String fileUrl) {
 
+    if (fileUrl == null || fileUrl.isEmpty()) {
+      return;
+    }
+
     String fileKey = extractFileKey(fileUrl);
     amazonS3.deleteObject(bucketName, fileKey);
   }
